requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
diskcache==5.6.3
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.311
langsmith==0.0.43
llama-index==0.8.41
llama_cpp_python==0.2.11
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numpy==1.26.0
openai==0.28.1
packaging==23.2
Panda3D==1.10.13
pandas==2.1.1
Pillow==10.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.17
yarl==1.9.2
